game: fix g_proxMineTimeout unit conversion (seconds to ms)
QL stores g_proxMineTimeout in seconds (default 20), not milliseconds.
The binary multiplies by 1000 in ProximityMine_Activate before adding
to level.time (verified at qagamex86.dll 0x1005b8b1: IMUL EAX, EAX, 0x3e8).

Without the conversion, prox mines were timing out after 20ms instead
of 20 seconds.

+++ code/game/g_missile.c
@@ -169,7 +169,9 @@ static void ProximityMine_Activate(gentity_t* ent) {
     float r;
 
     ent->think = ProximityMine_Explode;
-    ent->nextthink = level.time + g_proxMineTimeout.integer;
+    // [QL] g_proxMineTimeout is in SECONDS in QL (default 20s), binary does *1000 here
+    // (verified from QLDS qagamex86.dll: IMUL EAX, EAX, 0x3e8 at 0x1005b8b1)
+    ent->nextthink = level.time + g_proxMineTimeout.integer * 1000;
 
     ent->takedamage = qtrue;
     ent->health = 1;
